adapted to moved DataSinkElementFactory
