+++ bundles/org.palladiosimulator.servicelevelobjective.edp2/META-INF/MANIFEST.MF
@@ -18,7 +18,8 @@ Require-Bundle: org.apache.commons.math,
  org.palladiosimulator.edp2.visualization;bundle-version="1.0.0",
  org.palladiosimulator.edp2.visualization.jfreechart;bundle-version="1.0.0",
  org.palladiosimulator.measurementframework;bundle-version="1.0.0",
- org.palladiosimulator.servicelevelobjective;bundle-version="0.1.0"
+ org.palladiosimulator.servicelevelobjective;bundle-version="0.1.0",
+ org.palladiosimulator.edp2.datastream.ui;bundle-version="5.2.1"
 Bundle-RequiredExecutionEnvironment: JavaSE-17
 Bundle-ActivationPolicy: lazy
 Export-Package: org.palladiosimulator.servicelevelobjective.edp2,
